requirements.txt
Flask
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-flask
opentelemetry-instrumentation-wsgi
opentelemetry-instrumentation-requests
opentelemetry-instrumentation
opentelemetry-proto
